requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
#  AI_Legal_Assistant — requirements.txt
#
#  STEP 1 — Remove conflicting old package:
#       pip uninstall langchain-classic -y
#
#  STEP 2 — Install:
#       pip install -r requirements.txt
# ─────────────────────────────────────────────────────────────────────────────

# ── API framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.4

# ── JWT ───────────────────────────────────────────────────────────────────────
PyJWT==2.8.0

# ── LangChain stack ───────────────────────────────────────────────────────────
langchain==0.2.6
langchain-core==0.2.43
langchain-community==0.2.6
langchain-text-splitters==0.2.4
langchain-ollama==0.1.3
langchain-huggingface==0.0.3

# ── RAG / Vector store / Embeddings ──────────────────────────────────────────
faiss-cpu==1.8.0
sentence-transformers==3.0.1
pypdf==4.2.0

# ── Firebase ──────────────────────────────────────────────────────────────────
firebase-admin==6.5.0
